first take at PR comments
